Eliminating WebForms Controls from MacroContent and migrating the file from Web to Abstractions/Core project
