+++ src/Umbraco.Core/Macros/MacroContent.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace Umbraco.Web.Macros
+{
+    // represents the content of a macro
+    public class MacroContent
+    {
+        // gets or sets the text content
+        public string Text { get; set; }
+
+        // gets or sets the date the content was generated
+        public DateTime Date { get; set; } = DateTime.Now;
+
+        // a value indicating whether the content is empty
+        public bool IsEmpty => Text == null;
+
+        // gets an empty macro content
+        public static MacroContent Empty { get; } = new MacroContent();
+    }
+}
